+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased changes
+
+### Devices
+
+- Climate: add `group_address_active_state`, `group_address_command_value_state` and a `is_active` property.
+
 ## 0.18.6 NumericValue 2021-06-11
 
 ### Devices